perf: maddubs_i32→madd_i16 migration + bsums i16 + SIMD mins correction
Break the hidden maddubs_i32 2-instruction chain into explicit
maddubs_i16 (pmaddubsw) + madd_i16 (pmaddwd) calls across all
x86 kernels. q4k_dot bakes scale into madd_i16 — 2 instructions
per dot group instead of 3, matching llama.cpp AVX2 pattern.

- Migrate all kernels: q4k_dot, q4k_gemm_tile, q4k_fused_gemm,
  q6k_dot, bitnet_i2s, bitnet_i8dot (maddubs_i32 → explicit chain)
- Change bsums from i32 to i16 everywhere (quant kernels, dot kernels,
  FFI types, Rust matmul dispatch) — matches llama.cpp block_q8_K layout
- ARM q4k_dot: cvt_f16_f32 (fcvtl) replaces pow2 table for d/dmin
- ARM q4k_dot: addp_i16 (vpaddq_s16) for SIMD bsums pairing in mins_corr
- x86 q4k_dot: hadd_i16 (phaddw) for SIMD bsums pairing in row_mins

Results (Llama-3.2-3B Q4_K_M, 16 threads):
  prefill: 21.8 → 26.8 tok/s (+23%)
  decode:  7.4 → 7.9 tok/s (+7%)
  FFN GFLOPS: 148 → 198 (+34%)
  Gap to llama.cpp: 67%→82% prefill, 76%→81% decode

+++ src/inference/gemm_q4k.rs
@@ -15,7 +15,7 @@ pub(crate) struct BatchQ8K {
     pub qs_stride: usize,
     pub qs: Vec<i8>,
     pub d: Vec<f32>,
-    pub bsums: Vec<i32>,
+    pub bsums: Vec<i16>,
 }
 
 impl BatchQ8K {
@@ -26,7 +26,7 @@ impl BatchQ8K {
             n_tokens, dim, n_blocks, qs_stride,
             qs: vec![0i8; n_tokens * qs_stride],
             d: vec![0.0f32; n_tokens * n_blocks],
-            bsums: vec![0i32; n_tokens * n_blocks * 16],
+            bsums: vec![0i16; n_tokens * n_blocks * 16],
         }
     }
 
@@ -48,7 +48,7 @@ impl BatchQ8K {
     pub fn d_ptr(&self, t: usize) -> *const f32 {
         unsafe { self.d.as_ptr().add(t * self.n_blocks) }
     }
-    pub fn bsums_ptr(&self, t: usize) -> *const i32 {
+    pub fn bsums_ptr(&self, t: usize) -> *const i16 {
         unsafe { self.bsums.as_ptr().add(t * self.n_blocks * 16) }
     }
 }
@@ -280,7 +280,7 @@ pub(crate) fn q4k_fused_gemm_f32_mt(
         if start >= end { return; }
         let mut scores4 = [0.0f32; 4];
         let mut scratch = vec![0.0f32; 8];  // for kernel max-reduce (heap — must survive FFI)
-        let mut bs = vec![0i32; 16];        // for kernel bsums (heap — must survive FFI)
+        let mut bs = vec![0i16; 16];        // for kernel bsums (heap — must survive FFI)
         // Pre-cache f16→f32 weight scales ONCE (invariant across tokens)
         let mut da_w = [[0.0f32; MAX_BLOCKS]; 4];
         let mut dma_w = [[0.0f32; MAX_BLOCKS]; 4];
@@ -366,7 +366,7 @@ pub(crate) fn q4k_fused_silu_gemm_f32_mt(
         let mut g_scores = [0.0f32; 4];
         let mut u_scores = [0.0f32; 4];
         let mut scratch = vec![0.0f32; 8];
-        let mut bs_buf = vec![0i32; 16];
+        let mut bs_buf = vec![0i16; 16];
         let mut r = start;
         unsafe {
             while r + 4 <= end {
